+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.7.0] - 2024-05-17
+
+### Changed
+
+- Use ConfigKit with a metaconfig for all commands
+
+### Added
+
+- Gatecheck Config Encoder & Decoder
+
 ## [0.6.2] - 2024-05-15
 
 ### Changed